Reorganise superUsers
